refactor:OrderRequest 기능 Order로 통합 및 SSH 키 경로/권한 수정, 관련 코드 정리
