+++ http/orderRequest.http
@@ -1,36 +1,38 @@
 ### 구매 요청 생성
-POST http://localhost:8080/orders
-Content-Type: application/json
-Cookie: accessToken=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6InVzZXItMyIsImVtYWlsIjoidXNlckBjb2RlaXQuY29tIiwibmFtZSI6IuycoOyggCIsInJvbGUiOiJVU0VSIiwiaWF0IjoxNzUyMTkyNDE4LCJleHAiOjE3NTIxOTMzMTh9.Co6nDH_5mv1fddc-eNOei3MfuQ8MmVOLzdcQWnIApog
+# POST http://localhost:8080/orders
+# Content-Type: application/json
+# Cookie: accessToken=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6InVzZXItMSIsImVtYWlsIjoic3VwZXJfYWRtaW5AY29kZWl0LmNvbSIsIm5hbWUiOiLstZzqs6DqtIDrpqzsnpAiLCJyb2xlIjoiU1VQRVJfQURNSU4iLCJpYXQiOjE3NTIxOTM3NzEsImV4cCI6MTc1MjE5NDY3MX0.VbmsPDSKhvje758djPAngBPBGkLywUpjy7dhdAW_4No
 
-{
-  "cartItemIds": [1, 2, 3],
-  "adminMessage": "관리자에게 남길 메시지",
-  "requestMessage": "요청 메시지"
-}
+# {
+#   "cartItemIds": [1, 2, 3],
+#   "adminMessage": "관리자에게 남길 메시지",
+#   "requestMessage": "요청 메시지"
+# }
 
 ### 내 구매 요청 리스트 조회
-GET http://localhost:8080/orders
-Cookie: accessToken=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6InVzZXItMyIsImVtYWlsIjoidXNlckBjb2RlaXQuY29tIiwibmFtZSI6IuycoOyggCIsInJvbGUiOiJVU0VSIiwiaWF0IjoxNzUyMTkyNDE4LCJleHAiOjE3NTIxOTMzMTh9.Co6nDH_5mv1fddc-eNOei3MfuQ8MmVOLzdcQWnIApog
+# GET http://localhost:8080/orders
+# Cookie: accessToken=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6InVzZXItMyIsImVtYWlsIjoidXNlckBjb2RlaXQuY29tIiwibmFtZSI6IuycoOyggCIsInJvbGUiOiJVU0VSIiwiaWF0IjoxNzUyMTkyNDE4LCJleHAiOjE3NTIxOTMzMTh9.Co6nDH_5mv1fddc-eNOei3MfuQ8MmVOLzdcQWnIApog
 
 ### 구매 요청 상세 조회
-GET http://localhost:8080/orders/11
-Cookie: accessToken=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6InVzZXItMyIsImVtYWlsIjoidXNlckBjb2RlaXQuY29tIiwibmFtZSI6IuycoOyggCIsInJvbGUiOiJVU0VSIiwiaWF0IjoxNzUyMTkyNDE4LCJleHAiOjE3NTIxOTMzMTh9.Co6nDH_5mv1fddc-eNOei3MfuQ8MmVOLzdcQWnIApog
+# GET http://localhost:8080/orders/11
+# Cookie: accessToken=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6InVzZXItMyIsImVtYWlsIjoidXNlckBjb2RlaXQuY29tIiwibmFtZSI6IuycoOyggCIsInJvbGUiOiJVU0VSIiwiaWF0IjoxNzUyMTkyNDE4LCJleHAiOjE3NTIxOTMzMTh9.Co6nDH_5mv1fddc-eNOei3MfuQ8MmVOLzdcQWnIApog
 
 ### 구매 요청 취소
-PATCH http://localhost:8080/orders/11
-Content-Type: application/json
-Cookie: accessToken=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6InVzZXItMyIsImVtYWlsIjoidXNlckBjb2RlaXQuY29tIiwibmFtZSI6IuycoOyggCIsInJvbGUiOiJVU0VSIiwiaWF0IjoxNzUyMTkyNDE4LCJleHAiOjE3NTIxOTMzMTh9.Co6nDH_5mv1fddc-eNOei3MfuQ8MmVOLzdcQWnIApog
+# PATCH http://localhost:8080/orders/11
+# Content-Type: application/json
+# Cookie: accessToken=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6InVzZXItMyIsImVtYWlsIjoidXNlckBjb2RlaXQuY29tIiwibmFtZSI6IuycoOyggCIsInJvbGUiOiJVU0VSIiwiaWF0IjoxNzUyMTkyNDE4LCJleHAiOjE3NTIxOTMzMTh9.Co6nDH_5mv1fddc-eNOei3MfuQ8MmVOLzdcQWnIApog
 
-{
-  "status": "CANCELED"
-}
+# {
+#   "status": "CANCELED"
+# }
 
 ### 즉시 구매 (어드민 전용)
-POST http://localhost:8080/admin/orders/instant
-Content-Type: application/json
-Cookie: accessToken=어드민_토큰_필요
+# POST http://localhost:8080/admin/orders/instant
+# Content-Type: application/json
+# Cookie: accessToken=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6InVzZXItMSIsImVtYWlsIjoic3VwZXJfYWRtaW5AY29kZWl0LmNvbSIsIm5hbWUiOiLstZzqs6DqtIDrpqzsnpAiLCJyb2xlIjoiU1VQRVJfQURNSU4iLCJpYXQiOjE3NTIxOTM3NzEsImV4cCI6MTc1MjE5NDY3MX0.VbmsPDSKhvje758djPAngBPBGkLywUpjy7dhdAW_4No
 
-{
-  "cartItemIds": [1, 2]
-} 
+# {
+#   "cartItemIds": [1, 2]
+# }
+
+# OrderRequest 관련 HTTP 테스트가 Order 모델로 통합되어 주석 처리됨
