+++ src/dtos/orderRequest.dto.ts
@@ -1,13 +1,16 @@
-export type TCreateOrderRequestDto = {
-  adminMessage?: string;
-  requestMessage?: string;
-  cartItemIds: number[];
-};
+// export type TCreateOrderRequestDto = {
+//   adminMessage?: string;
+//   requestMessage?: string;
+//   cartItemIds: number[];
+// };
 
-export type TCreateInstantOrderRequestDto = {
-  cartItemIds: number[];
-};
+// export type TCreateInstantOrderRequestDto = {
+//   cartItemIds: number[];
+// };
 
-export type TUpdateOrderStatusDto = {
-  status: 'CANCELED';
-}; 
+// export type TUpdateOrderStatusDto = {
+//   status: 'CANCELED';
+// };
+
+// OrderRequest 관련 DTO가 Order 모델로 통합되어 주석 처리됨
+export {};
