+++ src/repositories/order.repository.ts
@@ -65,9 +65,162 @@ const updateOrder = async (
   });
 };
 
+// OrderRequest 관련 기능들 추가
+const createOrder = async (orderData: {
+  userId: string;
+  adminMessage?: string;
+  requestMessage?: string;
+  cartItemIds: number[];
+}, tx?: Prisma.TransactionClient) => {
+  const client = tx || prisma;
+
+  // 1. 카트 아이템들을 가져와서 receipt 생성
+  const cartItems = await client.cartItem.findMany({
+    where: {
+      id: { in: orderData.cartItemIds }
+    },
+    include: {
+      product: true
+    }
+  });
+
+  if (cartItems.length !== orderData.cartItemIds.length) {
+    throw new Error('일부 카트 아이템을 찾을 수 없습니다.');
+  }
+
+  // 2. 각 카트 아이템으로부터 receipt 생성
+  const receipts = await Promise.all(
+    cartItems.map(async (cartItem) => {
+      return await client.receipt.create({
+        data: {
+          productName: cartItem.product.name,
+          price: cartItem.product.price,
+          imageUrl: cartItem.product.imageUrl,
+          quantity: cartItem.quantity
+        }
+      });
+    })
+  );
+
+  // 3. 총 가격 계산
+  const totalPrice = receipts.reduce((sum, receipt) => {
+    return sum + (receipt.price * receipt.quantity);
+  }, 0);
+
+  // 4. 주문 생성
+  const order = await client.order.create({
+    data: {
+      userId: orderData.userId,
+      adminMessage: orderData.adminMessage,
+      requestMessage: orderData.requestMessage,
+      totalPrice: totalPrice,
+      status: 'PENDING'
+    }
+  });
+
+  // 5. OrderedItem 생성 (주문과 receipt 연결)
+  await Promise.all(
+    receipts.map(async (receipt) => {
+      return await client.orderedItem.create({
+        data: {
+          orderId: order.id,
+          receiptId: receipt.id
+        }
+      });
+    })
+  );
+
+  // 6. 생성된 주문 정보 반환
+  const result = await client.order.findUnique({
+    where: { id: order.id },
+    include: {
+      orderedItems: {
+        include: {
+          receipt: {
+            select: {
+              id: true,
+              productName: true,
+              price: true,
+              imageUrl: true,
+              quantity: true
+            }
+          }
+        }
+      }
+    }
+  });
+
+  if (!result) {
+    throw new Error('생성된 주문을 찾을 수 없습니다.');
+  }
+
+  return result;
+};
+
+const getOrdersByUserId = async (userId: string, tx?: Prisma.TransactionClient) => {
+  const client = tx || prisma;
+
+  return await client.order.findMany({
+    where: { userId },
+    include: {
+      orderedItems: {
+        include: {
+          receipt: {
+            select: {
+              id: true,
+              productName: true,
+              price: true,
+              imageUrl: true,
+              quantity: true
+            }
+          }
+        }
+      }
+    },
+    orderBy: { createdAt: 'desc' }
+  });
+};
+
+const updateOrderStatus = async (orderId: number, status: 'PENDING' | 'APPROVED' | 'REJECTED' | 'CANCELED', tx?: Prisma.TransactionClient) => {
+  const client = tx || prisma;
+
+  try {
+    const updatedOrder = await client.order.update({
+      where: { id: orderId },
+      data: { status },
+      include: {
+        orderedItems: {
+          include: {
+            receipt: {
+              select: {
+                id: true,
+                productName: true,
+                price: true,
+                imageUrl: true,
+                quantity: true
+              }
+            }
+          }
+        }
+      }
+    });
+
+    return updatedOrder;
+  } catch (error) {
+    if (error instanceof Prisma.PrismaClientKnownRequestError && error.code === 'P2025') {
+      throw new Error(`ID ${orderId}인 주문을 찾을 수 없습니다.`);
+    }
+    throw error;
+  }
+};
+
 export default {
   getOrders,
   getOrderByIdAndStatus,
   getOrderById,
   updateOrder,
+  // OrderRequest 관련 기능들
+  createOrder,
+  getOrdersByUserId,
+  updateOrderStatus,
 };
